refactor(cpp-boost-beast): share model pipeline between client and server
Move the direction-agnostic document pipeline, model lowering, parameter
serialization facts, dialect policy, and schema-IR emission from
CppBoostBeastClientCodegen into CppBoostBeastModelCodegen; extract
CppBoostBeastOperationFacts from the client template assembler; add
additionalEmbeddedTemplateDirs locator support and move shared
model/validation templates to cpp-boost-beast-common. Client generated
output is byte-identical; cppboostbeast suite (158 tests) green.
