+++ modules/openapi-generator/src/main/java/org/openapitools/codegen/languages/CppBoostBeastModelCodegen.java
@@ -99,6 +99,1298 @@ public abstract class CppBoostBeastModelCodegen extends AbstractCppCodegen {
     protected Map<String, List<String>> operationCallbacks = new HashMap<>();
     protected Map<String, List<String>> operationLinks = new HashMap<>();
 
+    protected CppBoostBeastModelCodegen() {
+        // Shared model/validation templates live in cpp-boost-beast-common so
+        // client and server generators resolve them from a single source.
+        additionalEmbeddedTemplateDirs = new ArrayList<>(List.of("cpp-boost-beast-common"));
+    }
+    public static final String EXPORT_MACRO = "exportMacro";
+    /** Policy for format metadata in composition branch matching.
+     *  Formats remain annotations and never affect branch match counts. */
+    protected String formatAssertionPolicy = "annotation";
+
+    /** Value type for the formatAssertion option. */
+    protected static final String FORMAT_ASSERTION_POLICY_ANNOTATION = "annotation";
+    /** Controls composition-branch validation during model decoding. */
+    protected boolean validateOnDecode = true;
+    /** Retains undeclared JSON object members in generated object models. */
+    protected boolean preserveAdditionalProperties = false;
+    private static final String X_CODEGEN_IS_RAW_BODY = "x-codegen-is-raw-body";
+    private static final String X_CODEGEN_IS_OPTIONAL_QUERY_PARAMETER =
+            "x-codegen-is-optional-query-parameter";
+    // Authoritative parameter serialization facts stamped by codegenParameterStyled().
+    private static final String X_CODEGEN_PARAM_STYLE = "x-codegen-param-style";
+    private static final String X_CODEGEN_PARAM_EXPLODE = "x-codegen-param-explode";
+    private static final String X_CODEGEN_PARAM_ALLOW_RESERVED =
+            "x-codegen-param-allow-reserved";
+    private static final String X_CODEGEN_PARAM_ALLOW_EMPTY_VALUE =
+            "x-codegen-param-allow-empty-value";
+    protected Map<String, String> componentSchemaIdsByName = Collections.emptyMap();
+    /** Starts an isolated state set for one generator invocation. */
+    protected void beginGeneration(OpenAPI openApi) {
+        sourceOpenApi = openApi;
+        variantModels = new HashSet<>();
+        resolvedAliasTypes = new HashMap<>();
+        composedKeywordsByModel = new HashMap<>();
+        compositionDescriptors = new LinkedHashMap<>();
+        compositionDescriptorSets = new LinkedHashMap<>();
+        webhookPreservation = new ArrayList<>();
+        operationCallbacks = new HashMap<>();
+        operationLinks = new HashMap<>();
+        allOfIntersections = new LinkedHashMap<>();
+        refreshComponentSchemaIds(openApi);
+
+    }
+
+    /**
+     * swagger-parser materializes the implicit root server as {@code /}, which
+     * is indistinguishable from a source-level {@code servers: [{url: /}]} in
+     * the model. Consult the raw document before server-precedence assembly.
+     */
+    protected boolean detectExplicitRootServers() {
+        String inputSpec = getInputSpec();
+        if (inputSpec == null || inputSpec.isEmpty()) {
+            // Programmatic OpenAPI instances have no parser-injected source.
+            return true;
+        }
+        try {
+            JsonNode document = Oas31RawSpecRecovery.readRawDocument(inputSpec);
+            return document != null && document.isObject() && document.has("servers");
+        } catch (Exception exception) {
+            throw new IllegalStateException(
+                    "Unable to inspect the source OpenAPI document for root servers", exception);
+        }
+    }
+
+    /**
+     * Returns the composition descriptor for the given schema name, or null
+     * if the schema is not composed or was not indexed.
+     */
+    public CompositionDescriptor getCompositionDescriptor(String schemaName) {
+        return compositionDescriptors.get(schemaName);
+    }
+
+    /**
+     * Returns an unmodifiable view of the full composition descriptor index.
+     */
+    public Map<String, CompositionDescriptor> getCompositionDescriptors() {
+        return Collections.unmodifiableMap(compositionDescriptors);
+    }
+
+    /**
+     * Returns every composition descriptor present on a schema, in keyword
+     * order: oneOf, anyOf, then allOf.
+     */
+    public List<CompositionDescriptor> getCompositionDescriptorsForSchema(String schemaName) {
+        return compositionDescriptorSets.getOrDefault(schemaName, Collections.emptyList());
+    }
+    /** Pinned OAS 3.1 Schema dialect (spec.openapis.org/oas/3.1/dialect/2024-11-10). */
+    public static final String OAS_31_DIALECT =
+            "https://spec.openapis.org/oas/3.1/dialect/2024-11-10";
+
+    /** OAS alias accepted only as the identifier for the same pinned revision. */
+    public static final String OAS_31_DIALECT_BASE_ALIAS =
+            "https://spec.openapis.org/oas/3.1/dialect/base";
+
+    /** Plain JSON Schema Draft 2020-12 core identifier (non-OAS dialect). */
+    public static final String DRAFT_2020_12 =
+            "https://json-schema.org/draft/2020-12/schema";
+
+    /** Classified effective schema dialect for an OpenAPI document. */
+    public enum OasDialect {
+        /** OAS 3.1 pinned dialect (or its base alias). */
+        OAS_31,
+        /** Plain JSON Schema Draft 2020-12 (not OAS-wrapped). */
+        DRAFT_2020_12_REC,
+        /** A dialect identifier not recognized by this program. */
+        UNRECOGNIZED,
+        /** No dialect declared (OAS 3.1 default applies for OAS 3.1 documents). */
+        UNSPECIFIED
+    }
+
+    /**
+     * Dialect resolution, normative-structure checks, and the exhaustive
+     * keyword-occurrence scanner live in {@link Oas31KeywordScanner};
+     * the delegates below keep this generator's public API stable for
+     * tests and templates.
+     */
+    public static OasDialect resolveEffectiveDialect(String jsonSchemaDialect, String rootSchema) {
+        return Oas31KeywordScanner.resolveEffectiveDialect(jsonSchemaDialect, rootSchema);
+    }
+
+    /** Resolve the effective dialect of an OpenAPI document from its declared knobs. */
+    public static OasDialect resolveDocumentDialect(OpenAPI openAPI) {
+        return Oas31KeywordScanner.resolveDocumentDialect(openAPI);
+    }
+
+    /** OAS 3 structural normative checks (see {@link Oas31KeywordScanner}). */
+    public List<String> validateNormativeOas3Structure(OpenAPI openAPI) {
+        return Oas31KeywordScanner.validateNormativeOas3Structure(openAPI);
+    }
+
+    /** Dialect/metaschema policy gate (see {@link Oas31KeywordScanner}). */
+    public List<String> validateDialectPolicy(OpenAPI openAPI) {
+        return Oas31KeywordScanner.validateDialectPolicy(openAPI);
+    }
+
+
+    /**
+     * Exhaustive schema-valued-position scanner (see {@link Oas31KeywordScanner}).
+     */
+    public Oas31KeywordScanner.KeywordOccurrenceLedger scanSchemaKeywordOccurrences(
+            OpenAPI openAPI) {
+        return Oas31KeywordScanner.scanSchemaKeywordOccurrences(openAPI);
+    }
+
+
+    /** Set of fail-closed required-vocabulary keywords for this document. */
+    public Set<String> failClosedKeywords(OpenAPI openAPI) {
+        return Oas31KeywordScanner.failClosedKeywords(openAPI);
+    }
+    /**
+     * Generator-specific normalizer that preserves composition structure
+     * (branch cardinality, null multiplicity, original keyword) for all
+     * oneOf/anyOf/anyOf-string-enum schemas. Set-equivalent simplification
+     * happens later in the generator's semantic analyzer (processComposedModel),
+     * never in the pre-descriptor normalizer.
+     */
+    public static class CppBoostBeastOpenAPINormalizer extends OpenAPINormalizer {
+        public CppBoostBeastOpenAPINormalizer(OpenAPI openAPI, Map<String, String> inputRules) {
+            super(openAPI, inputRules);
+        }
+
+        @Override
+        protected Schema processSimplifyAnyOf(Schema schema) {
+            if (schema.getAnyOf() != null && !schema.getAnyOf().isEmpty()) {
+                return schema;
+            }
+            return super.processSimplifyAnyOf(schema);
+        }
+
+        @Override
+        protected Schema processSimplifyOneOf(Schema schema) {
+            if (schema.getOneOf() != null && !schema.getOneOf().isEmpty()) {
+                return schema;
+            }
+            return super.processSimplifyOneOf(schema);
+        }
+
+        @Override
+        protected Schema processSimplifyAnyOfStringAndEnumString(Schema schema) {
+            if (schema.getAnyOf() != null && !schema.getAnyOf().isEmpty()) {
+                return schema;
+            }
+            return super.processSimplifyAnyOfStringAndEnumString(schema);
+        }
+
+        @Override
+        protected Schema processSimplifyOneOfEnum(Schema schema) {
+            if (schema.getOneOf() != null && !schema.getOneOf().isEmpty()) {
+                return schema;
+            }
+            return super.processSimplifyOneOfEnum(schema);
+        }
+
+        @Override
+        protected Schema processSimplifyAnyOfEnum(Schema schema) {
+            if (schema.getAnyOf() != null && !schema.getAnyOf().isEmpty()) {
+                return schema;
+            }
+            return super.processSimplifyAnyOfEnum(schema);
+        }
+    }
+    @Override
+    protected ImmutableMap.Builder<String, Lambda> addMustacheLambdas() {
+        return super.addMustacheLambdas()
+                .put("cppStringLiteral", (fragment, writer) -> writer.write(
+                        escapeCppStringContent(
+                                StringEscapeUtils.unescapeHtml4(fragment.execute()))));
+    }
+
+    @Override
+    public String escapeText(String input) {
+        return input == null ? null : escapeCppStringContent(input);
+    }
+    @Override
+    public String getterAndSetterCapitalize(String name) {
+        if (name == null || name.length() == 0) {
+            return name;
+        }
+
+        name = toVarName(name);
+
+        if (name.startsWith("_")) {
+            return "_" + camelize(name);
+        }
+
+        return camelize(name);
+    }
+    protected static boolean isSchemaValidationSupportingFile(SupportingFile file) {
+        String destination = file.getDestinationFilename();
+        return "Oas31SchemaRegistry.h".equals(destination)
+                || "schema_ir.generated.cpp".equals(destination)
+                || (destination.startsWith("schema_ir.generated.chunk")
+                        && destination.endsWith(".cpp"));
+    }
+    /**
+     * Location to write model files. You can use the modelPackage() as defined
+     * when the class is instantiated
+     */
+    @Override
+    public String modelFileFolder() {
+        return (outputFolder + "/model").replace("/", File.separator);
+    }
+
+    /**
+     * Location to write api files. You can use the apiPackage() as defined when
+     * the class is instantiated
+     */
+    @Override
+    public String apiFileFolder() {
+        return (outputFolder + "/api").replace("/", File.separator);
+    }
+
+    @Override
+    public String toModelImport(String name) {
+        if (importMapping.containsKey(name)) {
+            return importMapping.get(name);
+        } else {
+            return "#include \"" + name + ".h\"";
+        }
+    }
+
+    @Override
+    public CodegenModel fromModel(String name, Schema model) {
+        // Flatten allOf into a synthetic schema with intersected properties and
+        // unioned required names. Clearing allOf gives every property direct owned
+        // storage rather than generated inheritance.
+        Schema modelArg = model;
+        if (model != null && model.getAllOf() != null && !model.getAllOf().isEmpty()) {
+            AllOfIntersection intersection = allOfIntersections.get(
+                    toModelName(name));
+            if (intersection != null) {
+                // Check for unsatisfiable required properties / scalar conflicts
+                if (!intersection.isSatisfiable()) {
+                    throw new AllOfRequiredUnsatisfiableException(
+                            name, intersection.getUnsatisfiableReason());
+                }
+
+                Schema synthetic = Oas31CompositionLowering.buildSyntheticAllOfSchema(
+                        name, intersection);
+                // Copy top-level attributes from original model
+                if (model.getDiscriminator() != null) {
+                    synthetic.setDiscriminator(model.getDiscriminator());
+                }
+                if (Boolean.TRUE.equals(model.getNullable())) {
+                    synthetic.setNullable(true);
+                }
+                if (model.getDescription() != null) {
+                    synthetic.setDescription(model.getDescription());
+                }
+                if (model.getFormat() != null && intersection.getRootScalarType() != null) {
+                    synthetic.setFormat(model.getFormat());
+                }
+                // Optional impossible properties retain their API surface but
+                // reject any JSON object in which they are present.
+                if (!intersection.getOptionalImpossibleProperties().isEmpty()) {
+                    Map<String, Object> ext = synthetic.getExtensions();
+                    if (ext == null) {
+                        ext = new LinkedHashMap<>();
+                        synthetic.setExtensions(ext);
+                    }
+                    ext.put("x-cpp-optional-impossible-properties",
+                            new ArrayList<>(intersection.getOptionalImpossibleProperties()));
+                }
+                // Flat: allOf = null so super.fromModel sees no parent
+                synthetic.setAllOf(null);
+                modelArg = synthetic;
+            }
+        }
+
+        // Pre-check: The OpenAPI 3.1 parser converts anyOf [T, null] into
+        // {type: T, nullable: true} or {$ref: X, nullable: true}, consuming
+        // the anyOf list.  Detect these nullable schemas and produce the
+        // correct std::optional<T> type.
+        //
+        // For $ref schemas (normalised anyOf/oneOf [T, null] where T was a
+        // $ref), getTypeDeclaration resolves the target and returns the
+        // correct C++ type.  For arrays, getTypeDeclaration returns the
+        // container type (e.g. std::vector<...>) without optional wrapping,
+        // so we wrap it here.  Inline object schemas (type=object, no $ref)
+        // are full class models — they stay out of the alias precomputation
+        // because getTypeDeclaration would return the raw OAS type name
+        // "object" instead of the model name.  They are handled separately
+        // below via variant model registration.
+        boolean isNullableSchema = model != null
+            && Boolean.TRUE.equals(model.getNullable())
+            && (model.get$ref() != null
+                || (model.getType() != null && !"object".equals(model.getType())));
+        String preComputedNullUnionType = null;
+        if (isNullableSchema) {
+            // Resolve the type to its C++ type and wrap in std::optional
+            String innerType = getTypeDeclaration(model);
+            // getTypeDeclaration already returns std::optional<T> for nullable.
+            // Use it directly if it starts with std::optional<.
+            if (innerType.startsWith("std::optional<")) {
+                preComputedNullUnionType = innerType;
+            } else {
+                preComputedNullUnionType = "std::optional<" + innerType + ">";
+            }
+        } else if (model != null) {
+            // Also try the anyOf/oneOf path for cases where the parser
+            // preserved the composed schema structure.
+            preComputedNullUnionType = detectNullUnion(model, name);
+        }
+
+        CodegenModel codegenModel = super.fromModel(name, modelArg);
+        if (codegenModel == null) {
+            return null;
+        }
+
+        codegenModel.vendorExtensions.put(
+                "x-cpp-component-schema-id",
+                componentSchemaId(name, componentSchemaIdsByName));
+
+        // Post-check: Apply the pre-computed null union type if the default
+        // pipeline consumed the composed schemas.
+        if (preComputedNullUnionType != null) {
+            codegenModel.dataType = preComputedNullUnionType;
+            codegenModel.vendorExtensions.put("x-cpp-type", preComputedNullUnionType);
+            codegenModel.vendorExtensions.put("x-cpp-composed-keyword",
+                model.getAnyOf() != null ? "anyOf" : "oneOf");
+            codegenModel.vendorExtensions.put("x-cpp-is-alias", true);
+            codegenModel.vendorExtensions.put("x-cpp-is-optional", true);
+            // Force a model header/source so Gate A inventory and $ref users get
+            // `using NullableString = std::optional<std::string>;`. DefaultCodegen
+            // marks plain nullable primitives as isAlias and skips file emission.
+            codegenModel.isAlias = false;
+            resolvedAliasTypes.put(name, preComputedNullUnionType);
+            variantModels.add(name);
+        }
+
+        // Post-check: Inline nullable object schemas (type=object, nullable=true,
+        // no $ref) are full class models with properties — they cannot use the
+        // alias path. Register them as variant models so $ref references use value
+        // semantics (std::shared_ptr<NullableObject> → NullableObject) and tag
+        // the model as optional for correct null-value representation.
+        if (model != null && model.get$ref() == null
+                && "object".equals(model.getType())
+                && Boolean.TRUE.equals(model.getNullable())) {
+            variantModels.add(name);
+            codegenModel.vendorExtensions.put("x-cpp-is-optional", true);
+        }
+
+        Set<String> oldImports = codegenModel.imports;
+        codegenModel.imports = new HashSet<>();
+        for (String imp : oldImports) {
+            String newImp = toModelImport(imp);
+            if (!newImp.isEmpty()) {
+                codegenModel.imports.add(newImp);
+            }
+        }
+        // Every model header declares vector conversion helpers.
+        codegenModel.imports.add("#include <vector>");
+        if (preserveAdditionalProperties) {
+            codegenModel.imports.add("#include <map>");
+            codegenModel.imports.add("#include <string>");
+            codegenModel.imports.add("#include <utility>");
+            reserveExtraJsonPropertyIdentifiers(codegenModel);
+        }
+
+        // Fixed-const properties: OAS 3.1 `const`, single-value `enum`, or optional
+        // vendor extension `x-stainless-const`. Portable path is OAS `const` / single enum —
+        // vendor extensions are never required for correct encode/decode.
+        if (codegenModel.vars != null) {
+            Map<String, Schema> allProps = new LinkedHashMap<>();
+            if (model.getProperties() != null) {
+                allProps.putAll(model.getProperties());
+            }
+            if (model.getAllOf() != null && openAPI != null) {
+                for (Object parentObj : model.getAllOf()) {
+                    if (parentObj instanceof Schema) {
+                        Schema parentSchema = ModelUtils.getReferencedSchema(
+                                openAPI, (Schema) parentObj);
+                        if (parentSchema != null && parentSchema.getProperties() != null) {
+                            allProps.putAll(parentSchema.getProperties());
+                        }
+                    }
+                }
+            }
+            for (CodegenProperty var : codegenModel.vars) {
+                Object rawProp = allProps.get(var.baseName);
+                if (!(rawProp instanceof Schema)) {
+                    continue;
+                }
+                Schema varSchema = (Schema) rawProp;
+                boolean hasOasConst = varSchema.getConst() != null;
+                boolean hasSingleValueEnum = varSchema.getEnum() != null
+                        && varSchema.getEnum().size() == 1;
+                boolean hasStainlessConst = varSchema.getExtensions() != null
+                        && Boolean.TRUE.equals(varSchema.getExtensions().get("x-stainless-const"));
+                if (!(hasOasConst || hasSingleValueEnum || hasStainlessConst)) {
+                    continue;
+                }
+                String constRawValue = null;
+                if (varSchema.getConst() != null) {
+                    constRawValue = varSchema.getConst().toString();
+                } else if (varSchema.getEnum() != null && !varSchema.getEnum().isEmpty()) {
+                    constRawValue = varSchema.getEnum().get(0).toString();
+                }
+                if (constRawValue == null && var.example != null) {
+                    constRawValue = var.example;
+                }
+                if (constRawValue == null) {
+                    constRawValue = "std::string".equals(var.dataType) ? "" : "0";
+                }
+                String inlineValue;
+                boolean isStringConst = "std::string".equals(var.dataType)
+                        || "std::optional<std::string>".equals(var.dataType)
+                        || (var.isString && !var.isInteger && !var.isLong && !var.isNumber
+                        && !var.isBoolean);
+                if ("std::optional<std::string>".equals(var.dataType)) {
+                    inlineValue = "std::optional<std::string>{\""
+                            + escapeCppStringContent(constRawValue) + "\"}";
+                } else if (isStringConst || "std::string".equals(var.dataType)) {
+                    inlineValue = "\"" + escapeCppStringContent(constRawValue) + "\"";
+                } else {
+                    inlineValue = constRawValue;
+                }
+                // Neutral OAS-first flag used by templates.
+                var.vendorExtensions.put("x-cpp-const", true);
+                var.vendorExtensions.put("x-cpp-const-value", constRawValue);
+                var.vendorExtensions.put("x-cpp-const-inline-value", inlineValue);
+                // Mustache is truthy on key presence — only set when string-typed.
+                if (isStringConst || "std::string".equals(var.dataType)
+                        || "std::optional<std::string>".equals(var.dataType)) {
+                    var.vendorExtensions.put("x-cpp-const-is-string", true);
+                } else if (var.isBoolean || "bool".equals(var.dataType)
+                        || "std::optional<bool>".equals(var.dataType)) {
+                    var.vendorExtensions.put("x-cpp-const-is-boolean", true);
+                }
+                // Keep stainless keys as aliases so older template forks still work.
+                var.vendorExtensions.put("x-stainless-const", true);
+                var.vendorExtensions.put("x-stainless-const-value", constRawValue);
+                var.vendorExtensions.put("x-stainless-const-inline-value", inlineValue);
+            }
+        }
+
+        addContainerPropertyNames(codegenModel.vars);
+        return codegenModel;
+    }
+
+    @Override
+    public CodegenParameter fromParameter(Parameter parameter, Set<String> imports) {
+        CodegenParameter codegenParameter = super.fromParameter(parameter, imports);
+        // Preserve serialization facts for every parameter location.
+        codegenParameterStyled(parameter, codegenParameter);
+        if (!codegenParameter.isQueryParam) {
+            return codegenParameter;
+        }
+
+        if (!codegenParameter.required) {
+            codegenParameter.vendorExtensions.put(X_CODEGEN_IS_OPTIONAL_QUERY_PARAMETER, true);
+        }
+        return codegenParameter;
+    }
+
+    /**
+     * Records the OAS 3.1 serialization facts consumed by the C++ wire layer.
+     * Style defaults to form for query/cookie and simple for path/header. Explode
+     * defaults to true only for form. allowReserved is surfaced consistently;
+     * allowEmptyValue applies only to form-style query parameters.
+     */
+    protected void codegenParameterStyled(Parameter parameter,
+                                        CodegenParameter codegenParameter) {
+        String style = parameter.getStyle() == null
+                ? null : parameter.getStyle().toString();
+        if (style == null) {
+            if (codegenParameter.isQueryParam || codegenParameter.isCookieParam) {
+                style = "form";
+            } else {
+                style = "simple";   // path, header
+            }
+        }
+        Boolean explode = Boolean.TRUE.equals(parameter.getExplode());
+        if (parameter.getExplode() == null) {
+            explode = "form".equals(style);     // spec default
+        }
+        codegenParameter.vendorExtensions.put(X_CODEGEN_PARAM_STYLE, style);
+        codegenParameter.vendorExtensions.put(X_CODEGEN_PARAM_EXPLODE, explode);
+        codegenParameter.vendorExtensions.put(X_CODEGEN_PARAM_ALLOW_RESERVED,
+                Boolean.TRUE.equals(parameter.getAllowReserved()));
+        codegenParameter.vendorExtensions.put(X_CODEGEN_PARAM_ALLOW_EMPTY_VALUE,
+                Boolean.TRUE.equals(parameter.getAllowEmptyValue()));
+    }
+
+    protected String queryCollectionDelimiter(Parameter.StyleEnum style) {
+        if (style == Parameter.StyleEnum.SPACEDELIMITED) {
+            return "%20";
+        }
+        if (style == Parameter.StyleEnum.PIPEDELIMITED) {
+            return "%7C";
+        }
+        return ",";
+    }
+
+    protected void addContainerPropertyNames(List<CodegenProperty> properties) {
+        for (CodegenProperty property : properties) {
+            CodegenProperty item = property.items;
+            while (item != null) {
+                item.vendorExtensions.put("x-container-property-name", property.name);
+                item = item.items;
+            }
+        }
+    }
+
+    protected void reserveExtraJsonPropertyIdentifiers(CodegenModel codegenModel) {
+        Set<String> propertyAccessors = new HashSet<>();
+        Set<String> propertyMembers = new HashSet<>();
+        if (codegenModel.allVars != null) {
+            for (CodegenProperty property : codegenModel.allVars) {
+                if (property.getter != null) {
+                    propertyAccessors.add(property.getter);
+                }
+                if (property.setter != null) {
+                    propertyAccessors.add(property.setter);
+                }
+                if (property.name != null) {
+                    propertyMembers.add("m_" + property.name);
+                }
+            }
+        }
+
+        int maxSuffix = propertyAccessors.size() + propertyMembers.size() + 2;
+        for (int suffix = 1; suffix <= maxSuffix; suffix++) {
+            String suffixText = suffix == 1 ? "" : Integer.toString(suffix);
+            String getter = "getExtraJsonProperties" + suffixText;
+            String setter = "setExtraJsonProperties" + suffixText;
+            String member = "m_extraJsonProperties" + suffixText;
+            if (propertyAccessors.contains(getter) || propertyAccessors.contains(setter)
+                    || propertyMembers.contains(member)) {
+                continue;
+            }
+            codegenModel.vendorExtensions.put("x-cpp-extra-json-properties-getter", getter);
+            codegenModel.vendorExtensions.put("x-cpp-extra-json-properties-setter", setter);
+            codegenModel.vendorExtensions.put("x-cpp-extra-json-properties-member", member);
+            return;
+        }
+        throw new IllegalStateException("Unable to reserve C++ extra JSON property identifiers");
+    }
+
+    @Override
+    public String toModelFilename(String name) {
+        return toModelName(name);
+    }
+
+    @Override
+    public String toApiFilename(String name) {
+        return toApiName(name);
+    }
+    @Override
+    public String getTypeDeclaration(Schema p) {
+        // Handle inline oneOf/anyOf composed schemas (apply lowering rules directly)
+        if (ModelUtils.isComposedSchema(p) && (p.getOneOf() != null || p.getAnyOf() != null)) {
+            return lowerInlineComposedSchema(p);
+        }
+
+        String openAPIType = getSchemaType(p);
+
+        if (ModelUtils.isArraySchema(p)) {
+            // Use getItems() directly to handle both OpenAPI 3.0 and 3.1
+            Schema inner = p.getItems();
+            String arrayType;
+            if (inner != null) {
+                arrayType = getSchemaType(p) + "<" + getTypeDeclaration(inner) + ">";
+            } else {
+                arrayType = "std::vector<boost::json::value>";
+            }
+            // Nullable arrays must be wrapped in std::optional so null JSON
+            // values are representable. The array branch returns before the
+            // nullable fallback checks at the end of this method.
+            if (ModelUtils.isNullable(p)) {
+                return "std::optional<" + arrayType + ">";
+            }
+            return arrayType;
+        } else if (ModelUtils.isMapSchema(p)) {
+            Schema inner = ModelUtils.getAdditionalProperties(p);
+            String innerType = inner == null ? "boost::json::value" : getTypeDeclaration(inner);
+            String mapType = getSchemaType(p) + "<std::string, " + innerType + ">";
+            // Nullable maps must be wrapped in std::optional so null JSON
+            // values are representable. The map branch returns before the
+            // nullable fallback checks at the end of this method.
+            if (ModelUtils.isNullable(p)) {
+                return "std::optional<" + mapType + ">";
+            }
+            return mapType;
+        } else if (ModelUtils.isByteArraySchema(p)) {
+            return "std::string";
+        } else if (ModelUtils.isStringSchema(p)
+                || ModelUtils.isDateSchema(p)
+                || ModelUtils.isDateTimeSchema(p) || ModelUtils.isFileSchema(p)
+                || languageSpecificPrimitives.contains(openAPIType)
+                || typeMapping.containsKey(openAPIType)
+                || typeMapping.values().contains(openAPIType)) {
+            // Resolve through type mapping for scalar allOf: composed schemas
+            // return OAS raw types (e.g. "string") or mapped types (e.g.
+            // "std::string") depending on branch resolution path.
+            // Re-map if the value is already in the type mapping values.
+            String resolved = typeMapping.containsKey(openAPIType)
+                    ? typeMapping.get(openAPIType)
+                    : toModelName(openAPIType);
+            // OAS 3.0 nullable: true → std::optional<T>
+            if (ModelUtils.isNullable(p)) {
+                return "std::optional<" + resolved + ">";
+            }
+            return resolved;
+        } else if (ModelUtils.isNullType(p)) {
+            // Handle OpenAPI 3.1 null type
+            return "std::nullptr_t";
+        } else if (ModelUtils.isAnyType(p) || ModelUtils.isFreeFormObject(p, openAPI)) {
+            return "boost::json::value";
+        }
+
+        // OAS 3.0 nullable: true → std::optional<T>
+        if (ModelUtils.isNullable(p)) {
+            return "std::optional<" + openAPIType + ">";
+        }
+
+        // Variant models use value semantics (no shared_ptr wrapping)
+        if (variantModels.contains(openAPIType)) {
+            return openAPIType;
+        }
+
+        // Object references use shared ownership because circular-reference facts
+        // are unavailable when this declaration is computed. Variant aliases are
+        // handled above as value types.
+        return "std::shared_ptr<" + openAPIType + ">";
+    }
+
+    /**
+     * Resolves an inline oneOf/anyOf schema to its lowered C++ type by computing
+     * branch types and applying the same ordered lowering rules as model-level types.
+     */
+    protected String lowerInlineComposedSchema(Schema p) {
+        String composedKeyword;
+        List<Schema> children;
+        if (p.getOneOf() != null) {
+            children = p.getOneOf();
+            composedKeyword = "oneOf";
+        } else {
+            children = p.getAnyOf();
+            composedKeyword = "anyOf";
+        }
+
+        List<ComposedBranch> composedBranches = new ArrayList<>();
+        for (Schema child : children) {
+            // Compute the branch type using the full type declaration pipeline
+            // but strip shared_ptr for variant members (value semantics).
+            String childType = stripSharedPtr(getTypeDeclaration(child));
+            // Resolve $ref targets that are aliased to primitive types at the
+            // declaration point, before resolvedAliasTypes is available (it is
+            // populated during postProcessModels, which runs later). This handles
+            // inline schemas like CreateAssistantRequest_model = oneOf [string,
+            // $ref AssistantSupportedModels] where the target is anyOf [string,
+            // string-enum] → std::string, collapsing to just std::string.
+            Schema resolvedChild = child;
+            if (!childType.startsWith("std::") && !childType.startsWith("boost::")
+                    && !childType.startsWith("std::shared_ptr<")) {
+                Schema resolvedTarget = child.get$ref() != null && openAPI != null
+                        ? ModelUtils.getReferencedSchema(openAPI, child) : null;
+                if (resolvedTarget != null) {
+                    resolvedChild = resolvedTarget;
+                    String resolved = getTypeDeclaration(resolvedTarget);
+                    String stripped = stripSharedPtr(resolved);
+                    if (!stripped.equals(childType)) {
+                        childType = stripped;
+                    }
+                }
+            }
+            boolean isEnum = resolvedChild.getEnum() != null && !resolvedChild.getEnum().isEmpty();
+            boolean isStringLike = ModelUtils.isStringSchema(resolvedChild)
+                    || "std::string".equals(childType);
+            composedBranches.add(new ComposedBranch(childType, isEnum, isStringLike, -1));
+        }
+
+        // Deduplicate inside lowerComposedTypes so oneOf branch identity survives
+        // identical lowered C++ types.
+        return Oas31CompositionLowering.lowerComposedTypes(
+                composedBranches, composedKeyword, null, LOGGER::warn);
+    }
+    @Override
+    public CodegenProperty fromProperty(String name, Schema p, boolean required,
+                                        boolean schemaIsFromAdditionalProperties) {
+        CodegenProperty prop = super.fromProperty(name, p, required, schemaIsFromAdditionalProperties);
+        if (prop == null || p == null) {
+            return prop;
+        }
+        // Tag inline composed properties so templates can honor oneOf vs anyOf
+        // decode rules (exactly-one vs first-match) instead of always using
+        // the generic JsonValueConverter exactly-one path.
+        if (p.getOneOf() != null && !p.getOneOf().isEmpty()) {
+            prop.vendorExtensions.put("x-cpp-composed-keyword", "oneOf");
+            prop.vendorExtensions.put("x-cpp-is-oneof", true);
+        } else if (p.getAnyOf() != null && !p.getAnyOf().isEmpty()) {
+            prop.vendorExtensions.put("x-cpp-composed-keyword", "anyOf");
+            prop.vendorExtensions.put("x-cpp-is-anyof", true);
+        }
+        if (Oas31RawSpecRecovery.hasExplicitDefault(p)) {
+            String defaultValue = explicitScalarDefaultValue(prop, p);
+            if (defaultValue != null) {
+                prop.defaultValue = defaultValue;
+                prop.vendorExtensions.put("x-cpp-has-explicit-default", true);
+                prop.vendorExtensions.put(X_CPP_EXPLICIT_DEFAULT_SCALAR, defaultValue);
+                prop.vendorExtensions.put("x-cpp-default-is-null",
+                        "null".equals(Oas31RawSpecRecovery.defaultJsonOf(p)));
+            }
+        }
+        return prop;
+    }
+
+    protected String explicitScalarDefaultValue(CodegenProperty property, Schema schema) {
+        String json = Oas31RawSpecRecovery.defaultJsonOf(schema);
+        if (json == null) {
+            return null;
+        }
+
+        com.fasterxml.jackson.databind.JsonNode value;
+        try {
+            value = io.swagger.v3.core.util.Json31.mapper().readTree(json);
+        } catch (com.fasterxml.jackson.core.JsonProcessingException exception) {
+            throw new IllegalArgumentException(
+                    "Unable to parse default for property '" + property.baseName + "'", exception);
+        }
+        if (value == null || !value.isValueNode()) {
+            return null;
+        }
+
+        Object nullableInner = property.vendorExtensions.get(
+                "x-cpp-nullable-field-inner-type");
+        if (value.isNull()) {
+            if (nullableInner != null) {
+                return "NullableField<" + nullableInner + ">::makeDefaultNull()";
+            }
+            if (property.dataType != null
+                    && property.dataType.startsWith("std::optional<")) {
+                return "std::nullopt";
+            }
+            if ("std::nullptr_t".equals(property.dataType)) {
+                return "nullptr";
+            }
+            if ("boost::json::value".equals(property.dataType)) {
+                return "boost::json::value(nullptr)";
+            }
+            if (property.dataType != null
+                    && property.dataType.startsWith("std::shared_ptr<")) {
+                // A branch-local default:null is an annotation, not a model value.
+                // Ignore it rather than rejecting an otherwise legal schema.
+                return null;
+            }
+
+            throw new IllegalArgumentException(
+                    "JSON null default is not representable by C++ property '"
+                            + property.baseName + "' of type " + property.dataType);
+        }
+
+        String expression;
+        if (value.isTextual()) {
+            expression = "\"" + escapeCppStringContent(value.textValue()) + "\"";
+        } else if (value.isBoolean()) {
+            expression = value.booleanValue() ? "true" : "false";
+        } else if (value.isNumber()) {
+            expression = explicitNumericDefault(property, value.decimalValue());
+        } else {
+            return null;
+        }
+
+        if (nullableInner != null) {
+            return "NullableField<" + nullableInner + ">::makeDefaultValue("
+                    + expression + ")";
+        }
+        return expression;
+    }
+
+    protected static String explicitNumericDefault(
+            CodegenProperty property, java.math.BigDecimal value) {
+        if (property.isInteger || property.isLong) {
+            java.math.BigInteger integer;
+            try {
+                integer = value.toBigIntegerExact();
+            } catch (ArithmeticException exception) {
+                throw new IllegalArgumentException(
+                        "Non-integral default is not representable by integer property '"
+                                + property.baseName + "'", exception);
+            }
+            if (property.isLong || "std::int64_t".equals(property.dataType)) {
+                java.math.BigInteger min = java.math.BigInteger.valueOf(Long.MIN_VALUE);
+                java.math.BigInteger max = java.math.BigInteger.valueOf(Long.MAX_VALUE);
+                if (integer.compareTo(min) < 0 || integer.compareTo(max) > 0) {
+                    throw new IllegalArgumentException(
+                            "Default is outside int64 range for property '"
+                                    + property.baseName + "'");
+                }
+                if (integer.equals(min)) {
+                    return "std::int64_t{-9223372036854775807LL - 1LL}";
+                }
+                return "std::int64_t{" + integer + "LL}";
+            }
+            try {
+                return "std::int32_t{" + integer.intValueExact() + "}";
+            } catch (ArithmeticException exception) {
+                throw new IllegalArgumentException(
+                        "Default is outside int32 range for property '"
+                                + property.baseName + "'", exception);
+            }
+        }
+
+        String literal = value.toString();
+        boolean hasFloatingMarker = literal.indexOf('.') >= 0
+                || literal.indexOf('e') >= 0 || literal.indexOf('E') >= 0;
+        if (!hasFloatingMarker) {
+            literal += ".0";
+        }
+        if (property.isFloat || "float".equals(property.dataType)) {
+            float narrowed = value.floatValue();
+            if (!Float.isFinite(narrowed)
+                    || (value.signum() != 0 && narrowed == 0.0f)) {
+                throw new IllegalArgumentException(
+                        "Default is outside finite float range for property '"
+                                + property.baseName + "'");
+            }
+            return literal + "F";
+        }
+        double narrowed = value.doubleValue();
+        if (!Double.isFinite(narrowed)
+                || (value.signum() != 0 && narrowed == 0.0)) {
+            throw new IllegalArgumentException(
+                    "Default is outside finite double range for property '"
+                            + property.baseName + "'");
+        }
+        return literal;
+    }
+
+    @Override
+    public String toDefaultValue(Schema p) {
+        if (ModelUtils.isStringSchema(p)) {
+            if (p.getDefault() != null) {
+                return "\"" + escapeCppStringContent(p.getDefault().toString()) + "\"";
+            } else {
+                return "\"\"";
+            }
+        } else if (ModelUtils.isBooleanSchema(p)) {
+            if (p.getDefault() != null) {
+                return p.getDefault().toString();
+            } else {
+                return "false";
+            }
+        } else if (ModelUtils.isDateSchema(p)) {
+            if (p.getDefault() != null) {
+                return "\"" + escapeCppStringContent(p.getDefault().toString()) + "\"";
+            } else {
+                return "\"\"";
+            }
+        } else if (ModelUtils.isDateTimeSchema(p)) {
+            if (p.getDefault() != null) {
+                return "\"" + escapeCppStringContent(p.getDefault().toString()) + "\"";
+            } else {
+                return "\"\"";
+            }
+        } else if (ModelUtils.isNumberSchema(p)) {
+            if (ModelUtils.isFloatSchema(p)) { // float
+                if (p.getDefault() != null) {
+                    return p.getDefault().toString() + "f";
+                } else {
+                    return "0.0f";
+                }
+            } else { // double
+                if (p.getDefault() != null) {
+                    return p.getDefault().toString();
+                } else {
+                    return "0.0";
+                }
+            }
+        } else if (ModelUtils.isIntegerSchema(p)) {
+            if (ModelUtils.isLongSchema(p)) { // long
+                if (p.getDefault() != null) {
+                    return p.getDefault().toString() + "L";
+                } else {
+                    return "0L";
+                }
+            } else { // integer
+                if (p.getDefault() != null) {
+                    return p.getDefault().toString();
+                } else {
+                    return "0";
+                }
+            }
+        } else if (ModelUtils.isByteArraySchema(p)) {
+            if (p.getDefault() != null) {
+                return "\"" + escapeCppStringContent(p.getDefault().toString()) + "\"";
+            } else {
+                return "\"\"";
+            }
+        } else if (ModelUtils.isMapSchema(p)) {
+            Schema inner = ModelUtils.getAdditionalProperties(p);
+            String innerType = inner == null ? "boost::json::value" : getTypeDeclaration(inner);
+            return "std::map<std::string, " + innerType + ">()";
+        } else if (ModelUtils.isArraySchema(p)) {
+            // Use getItems() directly to handle OpenAPI 3.1 JsonSchema
+            Schema inner = p.getItems();
+            String innerType = inner != null ? getTypeDeclaration(inner) : "boost::json::value";
+            return "std::vector<" + innerType + ">()";
+        } else if (!StringUtils.isEmpty(p.get$ref())) {
+            String refName = toModelName(ModelUtils.getSimpleRef(p.get$ref()));
+            if (variantModels.contains(refName)) {
+                return refName + "()";
+            }
+            return "std::make_shared<" + refName + ">()";
+        } else if (ModelUtils.isNullType(p)) {
+            return "nullptr";
+        } else if (ModelUtils.isAnyType(p) || ModelUtils.isFreeFormObject(p, openAPI)) {
+            return "boost::json::value()";
+        }
+
+        return "nullptr";
+    }
+    
+    @Override
+    public String toDefaultValue(CodegenProperty codegenProperty, Schema schema) {
+        if (codegenProperty != null) {
+            if (codegenProperty.dataType != null && codegenProperty.dataType.startsWith("std::shared_ptr<")) {
+                return "nullptr";
+            }
+            if ("boost::json::value".equals(codegenProperty.dataType)) {
+                return "boost::json::value()";
+            }
+            Schema referenceSchema = Oas31CompositionLowering.referenceSchemaOf(schema);
+            if (referenceSchema != null && referenceSchema != schema
+                    && schema.getDefault() == null) {
+                Schema referencedTarget = ModelUtils.getReferencedSchema(openAPI, referenceSchema);
+                if (referencedTarget != null && referencedTarget != referenceSchema
+                        && codegenProperty.dataType != null
+                        && codegenProperty.dataType.equals(getTypeDeclaration(referencedTarget))) {
+                    return toDefaultValue(referencedTarget);
+                }
+            }
+        }
+        return super.toDefaultValue(codegenProperty, schema);
+    }
+
+    @Override
+    public void setParameterEncodingValues(CodegenParameter codegenParameter, MediaType mediaType) {
+        super.setParameterEncodingValues(codegenParameter, mediaType);
+        // Detect Encoding Object headers that cannot be propagated to
+        // multipart parts. When an Encoding Object specifies headers,
+        // emit a diagnostic instead of silently dropping them.
+        if (codegenParameter.isFormParam && mediaType != null
+                && mediaType.getEncoding() != null) {
+            io.swagger.v3.oas.models.media.Encoding encoding =
+                    mediaType.getEncoding().get(codegenParameter.baseName);
+            if (encoding != null && encoding.getHeaders() != null
+                    && !encoding.getHeaders().isEmpty()) {
+                LOGGER.warn("Encoding Object on form parameter '{}' specifies {} header(s) "
+                        + "that are not propagated to the multipart part. "
+                        + "Generated code uses only the contentType field. "
+                        + "Header keys: {}",
+                        codegenParameter.baseName,
+                        encoding.getHeaders().size(),
+                        encoding.getHeaders().keySet());
+            }
+        }
+    }
+    @Override
+    public void postProcessParameter(CodegenParameter parameter) {
+        super.postProcessParameter(parameter);
+
+        boolean isPrimitiveType = parameter.isPrimitiveType == Boolean.TRUE;
+        boolean isArray = parameter.isArray == Boolean.TRUE;
+        boolean isMap = parameter.isMap == Boolean.TRUE;
+        boolean isString = parameter.isString == Boolean.TRUE;
+        parameter.vendorExtensions.put(X_CODEGEN_IS_RAW_BODY,
+                isPrimitiveType || isString || parameter.isByteArray || parameter.isBinary
+                        || "std::string".equals(parameter.dataType));
+
+        if (!isPrimitiveType && !isArray && !isMap && !isString && !parameter.dataType.startsWith("std::shared_ptr")
+                && !"boost::json::value".equals(parameter.dataType)
+                && !"std::nullptr_t".equals(parameter.dataType)
+                && !parameter.dataType.startsWith("std::variant<")
+                && !parameter.dataType.startsWith("std::optional<")
+                && !"std::monostate".equals(parameter.dataType)) {
+            // Wrap non-primitive types in shared_ptr, unless:
+            // - The type is a variant/optional model (value semantics)
+            // - The type is a known variant model name from composed schemas
+            if (!variantModels.contains(parameter.dataType)) {
+                parameter.dataType = "std::shared_ptr<" + parameter.dataType + ">";
+                parameter.defaultValue = "std::make_shared<" + parameter.dataType + ">()";
+            }
+        }
+
+        // Post-hoc unwrap: if the type ended up as std::shared_ptr<VariantModel>,
+        // strip the shared_ptr wrapper (value semantics for variant types).
+        if (parameter.dataType != null && parameter.dataType.startsWith("std::shared_ptr<")
+                && parameter.dataType.endsWith(">")) {
+            String innerType = parameter.dataType.substring(16, parameter.dataType.length() - 1);
+            if (variantModels.contains(innerType)) {
+                parameter.dataType = innerType;
+                parameter.defaultValue = null;
+            }
+        }
+
+        // For form params, validate that encoding style/explode combinations
+        // are representable in multipart/form-data. Only form-style is supported
+        // for multipart (space-delimited, pipe-delimited, and deep-object styles
+        // are not representable). Fail closed with a targeted diagnostic.
+        if (parameter.isFormParam) {
+            if (Boolean.TRUE.equals(parameter.isSpaceDelimited)) {
+                throw new UnsupportedSchemaAssertionException(
+                        parameter.baseName,
+                        "encoding-style");
+            }
+            if (Boolean.TRUE.equals(parameter.isPipeDelimited)) {
+                throw new UnsupportedSchemaAssertionException(
+                        parameter.baseName,
+                        "encoding-style");
+            }
+            if (Boolean.TRUE.equals(parameter.isDeepObject)) {
+                throw new UnsupportedSchemaAssertionException(
+                        parameter.baseName,
+                        "encoding-style");
+            }
+        }
+
+        // Tag variant form params for branch-aware multipart serialization.
+        // When a form parameter's type is a variant, the template uses
+        // addVariantFormParameter to dispatch binary branches as file parts
+        // and object branches as JSON parts.
+        // Only set for actual std::variant types, not for models that alias
+        // to primitive types (e.g., VideoModel → std::string), which would
+        // cause instantiation of addVariantFormParameter<std::string> and
+        // an invalid std::visit call on a non-variant type.
+        boolean isVariantParam = false;
+        if (parameter.isFormParam && parameter.dataType != null) {
+            if (parameter.dataType.startsWith("std::variant<")) {
+                isVariantParam = true;
+            } else if (variantModels.contains(parameter.dataType)) {
+                String resolved = resolveThroughAliases(parameter.dataType);
+                if (resolved != null && resolved.startsWith("std::variant<")) {
+                    isVariantParam = true;
+                }
+            }
+        }
+        if (isVariantParam) {
+            parameter.vendorExtensions.put("x-codegen-is-variant-form-param", true);
+        }
+    }
+    @Override
+    public String getSchemaType(Schema p) {
+        // Non-standard format (NOT core OAS vocabulary). Documented generator
+        // convenience for corpora that use Unix-epoch integer timestamps.
+        // Disable by not using format: unixtime in the source document.
+        if (p != null && "unixtime".equals(p.getFormat())) {
+            return "int64_t";
+        }
+        String openAPIType = super.getSchemaType(p);
+        String type = null;
+        String modelName;
+        if (typeMapping.containsKey(openAPIType)) {
+            type = typeMapping.get(openAPIType);
+        } else {
+            type = openAPIType;
+        }
+
+        modelName = toModelName(type);
+        return modelName;
+    }
+    @Override
+    public void updateCodegenPropertyEnum(CodegenProperty var) {
+        // Remove prefix added by DefaultCodegen
+        String originalDefaultValue = var.defaultValue;
+        super.updateCodegenPropertyEnum(var);
+        var.defaultValue = originalDefaultValue;
+    }
+    protected void refreshComponentSchemaIds(OpenAPI openApi) {
+        if (openApi == null || openApi.getComponents() == null
+                || openApi.getComponents().getSchemas() == null) {
+            componentSchemaIdsByName = Collections.emptyMap();
+            return;
+        }
+        componentSchemaIdsByName = componentSchemaIds(
+                openApi.getComponents().getSchemas().keySet());
+    }
+    @Override
+    public Map<String, Object> postProcessSupportingFileData(Map<String, Object> objs) {
+        Map<String, Object> processed = super.postProcessSupportingFileData(objs);
+        if (!validateOnDecode) {
+            return processed;
+        }
+        // Model processing can replace inline branch schema objects after the
+        // initial recovery pass; refresh the emitted graph from the raw spec.
+        Oas31RawSpecRecovery.recoverPristineLiterals(openAPI, getInputSpec());
+        refreshComponentSchemaIds(openAPI);
+        Oas31SchemaIrEmitter emitter = new Oas31SchemaIrEmitter(
+                openAPI, compositionDescriptors, additionalProperties(), componentSchemaIdsByName);
+        Map<String, Object> produced = emitter.produce(processed);
+        supportingFiles.removeIf(file -> {
+            String destination = file.getDestinationFilename();
+            return destination.startsWith("schema_ir.generated.chunk")
+                    && destination.endsWith(".cpp");
+        });
+        int chunkCount = ((Number) produced.get("oas31SchemaIrChunkCount")).intValue();
+        for (int chunk = 0; chunk < chunkCount; chunk++) {
+            supportingFiles.add(new SupportingFile(
+                    Oas31SchemaIrEmitter.schemaIrChunkTemplate(chunk),
+                    "model", Oas31SchemaIrEmitter.schemaIrChunkFilename(chunk)));
+        }
+        return produced;
+    }
+    @Override
+    public void preprocessOpenAPI(OpenAPI openAPI) {
+        beginGeneration(openAPI);
+        hasExplicitRootServers = detectExplicitRootServers();
+
+        List<String> policyDiagnostics = validateDialectPolicy(openAPI);
+        if (!policyDiagnostics.isEmpty()) {
+            throw new IllegalArgumentException(String.join("; ", policyDiagnostics));
+        }
+        super.preprocessOpenAPI(openAPI);
+        // Webhooks are inbound-only metadata for a client generator. Upstream
+        // folds them into the API map under the same fallback classname as path
+        // operations, which can replace the path API. Preserve their metadata,
+        // then remove them so outbound paths still generate; no listener is emitted.
+        if (openAPI.getWebhooks() != null && !openAPI.getWebhooks().isEmpty()) {
+            for (Map.Entry<String, PathItem> e : openAPI.getWebhooks().entrySet()) {
+                PathItem item = e.getValue();
+                List<String> methods = new ArrayList<>();
+                if (item.getGet() != null) methods.add("GET " + idOf(item.getGet()));
+                if (item.getPut() != null) methods.add("PUT " + idOf(item.getPut()));
+                if (item.getPost() != null) methods.add("POST " + idOf(item.getPost()));
+                if (item.getDelete() != null) methods.add("DELETE " + idOf(item.getDelete()));
+                if (item.getPatch() != null) methods.add("PATCH " + idOf(item.getPatch()));
+                if (item.getHead() != null) methods.add("HEAD " + idOf(item.getHead()));
+                if (item.getOptions() != null) methods.add("OPTIONS " + idOf(item.getOptions()));
+                if (item.getTrace() != null) methods.add("TRACE " + idOf(item.getTrace()));
+                webhookPreservation.add(e.getKey()
+                        + "[" + String.join(", ", methods) + "]");
+            }
+            openAPI.setWebhooks(null);
+        }
+        // Capture callback and response-link names for generated API comments.
+        captureOperationMetadata(openAPI);
+        // Recover prefixItems dropped when the shared OAS 3.1 normalizer
+        // converts a type-array JsonSchema to ArraySchema. This must precede
+        // descriptor scanning so child schemas retain the pristine value.
+        Oas31RawSpecRecovery.restoreNormalizerDroppedPrefixItems(openAPI, getInputSpec());
+        Oas31RawSpecRecovery.recoverPristineLiterals(openAPI, getInputSpec());
+        // Populate variantModels and build composition descriptors before
+        // model processing begins so that getTypeDeclaration can resolve $ref
+        // to composed models as value types and branch semantics are captured
+        // before fromModel consumes composed schemas.
+        Map<String, Schema> schemas = openAPI.getComponents() != null
+                ? openAPI.getComponents().getSchemas() : null;
+        if (schemas != null) {
+            // Build descriptor index: must happen after inline model resolver
+            // flattening so all inline schemas have been extracted to component
+            // references with stable $ref targets.
+            for (Map.Entry<String, Schema> entry : schemas.entrySet()) {
+                String schemaName = entry.getKey();
+                Schema schema = entry.getValue();
+                List<CompositionDescriptor> descriptors =
+                        Oas31CompositionLowering.buildCompositionDescriptors(
+                                schemaName, schema, openAPI, schemas);
+                if (!descriptors.isEmpty()) {
+                    String modelName = toModelName(schemaName);
+                    // The primary descriptor drives representation lowering;
+                    // retain and validate every composition keyword separately.
+                    compositionDescriptors.put(modelName, descriptors.get(0));
+                    compositionDescriptorSets.put(modelName, Collections.unmodifiableList(
+                            new ArrayList<>(descriptors)));
+                    for (CompositionDescriptor descriptor : descriptors) {
+                        Oas31CompositionLowering.validateDescriptorAssertions(descriptor);
+                    }
+                }
+                // allOf affects object storage even when oneOf or anyOf selects
+                // the public representation.
+                if (schema.getAllOf() != null && !schema.getAllOf().isEmpty()) {
+                    AllOfIntersection intersection =
+                            Oas31CompositionLowering.computeAllOfIntersection(
+                                    schemaName, schema, openAPI, schemas, new HashSet<>());
+                    if (intersection != null) {
+                        allOfIntersections.put(toModelName(schemaName), intersection);
+                    }
+                }
+                if ((schema.getOneOf() != null && !schema.getOneOf().isEmpty())
+                        || (schema.getAnyOf() != null && !schema.getAnyOf().isEmpty())) {
+                    variantModels.add(schemaName);
+                }
+            }
+        }
+}
+
+    /** Shared namespace/validation options common to client and server generators. */
+    protected void applySharedCppOptions() {
+        String modelNamespace = modelPackage.replaceAll("\\.", "::");
+        additionalProperties.put("modelNamespaceDeclarations", modelPackage.split("\\."));
+        additionalProperties.put("modelNamespace", modelNamespace);
+        additionalProperties.put("schemaValidationNamespace",
+                modelNamespace + "::detail::schema_validation");
+        additionalProperties.put("schemaValidationHeaderGuardPrefix",
+                modelPackage.replaceAll("[^A-Za-z0-9]", "_").toUpperCase(Locale.ROOT));
+        additionalProperties.put("apiHeaderGuardPrefix",
+                apiPackage.replaceAll("[^A-Za-z0-9]", "_").toUpperCase(Locale.ROOT));
+        additionalProperties.put("apiNamespaceDeclarations", apiPackage.split("\\."));
+        additionalProperties.put("apiNamespace", apiPackage.replaceAll("\\.", "::"));
+
+        if (additionalProperties.containsKey("formatAssertionPolicy")) {
+            String policy = additionalProperties.get("formatAssertionPolicy")
+                    .toString().trim().toLowerCase(Locale.ROOT);
+            if (!FORMAT_ASSERTION_POLICY_ANNOTATION.equals(policy)) {
+                throw new IllegalArgumentException(
+                        "formatAssertionPolicy supports only 'annotation'; "
+                                + "format assertions are not implemented");
+            }
+        }
+        formatAssertionPolicy = FORMAT_ASSERTION_POLICY_ANNOTATION;
+        additionalProperties.put("formatAssertionPolicy", formatAssertionPolicy);
+
+        if (additionalProperties.containsKey("compileWithValidation")) {
+            Object raw = additionalProperties.get("compileWithValidation");
+            if (raw instanceof Boolean) {
+                validateOnDecode = (Boolean) raw;
+            } else {
+                validateOnDecode = Boolean.parseBoolean(raw.toString().trim());
+            }
+        }
+        additionalProperties.put("validateOnDecode", validateOnDecode);
+        additionalProperties.put("compileWithValidation", validateOnDecode);
+        if (!validateOnDecode) {
+            supportingFiles.removeIf(CppBoostBeastModelCodegen::isSchemaValidationSupportingFile);
+        }
+        preserveAdditionalProperties = false;
+        if (additionalProperties.containsKey("preserveAdditionalProperties")) {
+            Object raw = additionalProperties.get("preserveAdditionalProperties");
+            if (raw instanceof Boolean) {
+                preserveAdditionalProperties = (Boolean) raw;
+            } else {
+                String value = raw.toString().trim();
+                if (!"true".equalsIgnoreCase(value) && !"false".equalsIgnoreCase(value)) {
+                    throw new IllegalArgumentException(
+                            "preserveAdditionalProperties must be true or false: " + value);
+                }
+                preserveAdditionalProperties = Boolean.parseBoolean(value);
+            }
+        }
+        additionalProperties.put("preserveAdditionalProperties", preserveAdditionalProperties);
+        if (additionalProperties.containsKey("tolerateNonNullableNulls")) {
+            Object raw = additionalProperties.get("tolerateNonNullableNulls");
+            if (raw instanceof Boolean) {
+                tolerateNonNullableNulls = (Boolean) raw;
+            } else {
+                tolerateNonNullableNulls = Boolean.parseBoolean(raw.toString().trim());
+            }
+        }
+        additionalProperties.put("tolerateNonNullableNulls", tolerateNonNullableNulls);
+    }
+
+
     protected void captureOperationMetadata(OpenAPI openAPI) {
         operationCallbacks.clear();
         operationLinks.clear();
@@ -222,6 +1514,18 @@ public abstract class CppBoostBeastModelCodegen extends AbstractCppCodegen {
             }
         }
 
+        // Stamp every model with its component schema IR id after upstream
+        // model updates complete (mirrors the pre-refactor client pipeline).
+        refreshComponentSchemaIds(openAPI);
+        for (Map.Entry<String, ModelsMap> entry : objs.entrySet()) {
+            for (ModelMap modelMap : entry.getValue().getModels()) {
+                CodegenModel model = modelMap.getModel();
+                String schemaName = model.schemaName != null
+                        ? model.schemaName : entry.getKey();
+                model.vendorExtensions.put("x-cpp-component-schema-id",
+                        componentSchemaId(schemaName, componentSchemaIdsByName));
+            }
+        }
         return objs;
     }
