+++ modules/openapi-generator/src/main/java/org/openapitools/codegen/languages/CppBoostBeastOperationFacts.java
@@ -1,0 +1,133 @@
+package org.openapitools.codegen.languages;
+
+import io.swagger.v3.oas.models.OpenAPI;
+import io.swagger.v3.oas.models.PathItem;
+import io.swagger.v3.oas.models.security.SecurityRequirement;
+import io.swagger.v3.oas.models.security.SecurityScheme;
+import io.swagger.v3.oas.models.servers.Server;
+import org.openapitools.codegen.CodegenOperation;
+
+import java.util.ArrayList;
+import java.util.LinkedHashMap;
+import java.util.List;
+import java.util.Map;
+
+/**
+ * Direction-agnostic operation facts shared by the Boost.Beast client and
+ * server template assemblers: raw-operation lookup, effective security
+ * groups, and server-list classification.
+ */
+final class CppBoostBeastOperationFacts {
+    private CppBoostBeastOperationFacts() {
+    }
+
+    /** True when the list is exactly swagger-parser's implicit root default
+     *  (a single Server with url "/") and the raw source omitted {@code servers}. */
+    static boolean isParserDefaultServerList(List<Server> servers) {
+        return servers != null && servers.size() == 1
+                && "/".equals(servers.get(0).getUrl());
+    }
+
+    /** The operation's effective security requirements as template-ready
+     *  groups. Each group is an OR alternative containing AND-required scheme
+     *  maps. An empty group is anonymous access; operation {@code security: []}
+     *  clears inherited requirements. */
+    static List<List<Map<String, Object>>> effectiveSecurityGroups(
+            OpenAPI document, CodegenOperation op) {
+        List<List<Map<String, Object>>> groups = new ArrayList<>();
+        List<SecurityRequirement> requirements = null;
+        io.swagger.v3.oas.models.Operation raw = operationFor(document, op);
+        if (raw != null && raw.getSecurity() != null) {
+            requirements = raw.getSecurity();        // includes `[]` clears
+        } else if (document != null
+                && document.getSecurity() != null) {
+            requirements = document.getSecurity();
+        }
+        if (requirements == null) {
+            return groups;                            // no security declared
+        }
+        Map<String, SecurityScheme> schemes = document != null
+                && document.getComponents() != null
+                ? document.getComponents().getSecuritySchemes()
+                : null;
+        for (SecurityRequirement req : requirements) {
+            List<Map<String, Object>> ands = new ArrayList<>();
+            if (req != null) {
+                for (Map.Entry<String, List<String>> e : req.entrySet()) {
+                    SecurityScheme scheme = schemes == null
+                            ? null : schemes.get(e.getKey());
+                    Map<String, Object> use = new LinkedHashMap<>();
+                    use.put("name", cppString(e.getKey()));
+                    use.put("type", cppString(scheme == null || scheme.getType() == null
+                            ? "unknown" : scheme.getType().toString()));
+                    if (scheme != null && scheme.getType() == SecurityScheme.Type.APIKEY) {
+                        use.put("in", cppString(scheme.getIn() == null ? "header"
+                                : scheme.getIn().toString()));
+                        use.put("paramName", cppString(scheme.getName() == null
+                                ? "" : scheme.getName()));
+                    } else {
+                        use.put("in", "");
+                        use.put("paramName", "");
+                    }
+                    use.put("httpScheme", cppString(scheme != null
+                            && scheme.getType() == SecurityScheme.Type.HTTP
+                            && scheme.getScheme() != null
+                            ? scheme.getScheme() : ""));
+                    List<String> scopes = e.getValue() == null
+                            ? new ArrayList<String>() : e.getValue();
+                    use.put("scopes", scopes);
+                    use.put("scopesRendered", scopes.isEmpty() ? null
+                            : scopes.stream()
+                                    .map(s -> "\"" + cppString(s) + "\"")
+                                    .collect(java.util.stream.Collectors
+                                            .joining(", ")));
+                    ands.add(use);
+                }
+            }
+            groups.add(ands);                          // empty ands = {}
+        }
+        return groups;
+    }
+
+    /** The raw Operation behind a CodegenOperation (PathItem-method lookup). */
+    static io.swagger.v3.oas.models.Operation operationFor(
+            OpenAPI document, CodegenOperation op) {
+        if (document == null || document.getPaths() == null) {
+            return null;
+        }
+        PathItem item = document.getPaths().get(op.path);
+        if (item == null) {
+            return null;
+        }
+        if ("GET".equals(op.httpMethod)) {
+            return item.getGet();
+        }
+        if ("PUT".equals(op.httpMethod)) {
+            return item.getPut();
+        }
+        if ("POST".equals(op.httpMethod)) {
+            return item.getPost();
+        }
+        if ("DELETE".equals(op.httpMethod)) {
+            return item.getDelete();
+        }
+        if ("OPTIONS".equals(op.httpMethod)) {
+            return item.getOptions();
+        }
+        if ("HEAD".equals(op.httpMethod)) {
+            return item.getHead();
+        }
+        if ("PATCH".equals(op.httpMethod)) {
+            return item.getPatch();
+        }
+        if ("TRACE".equals(op.httpMethod)) {
+            return item.getTrace();
+        }
+        return null;
+    }
+
+    private static String cppString(String value) {
+        return CppBoostBeastModelCodegen.escapeCppStringContent(
+                value == null ? "" : value);
+    }
+}
