+++ modules/openapi-generator/src/main/resources/cpp-boost-beast-common/NullableField.h.mustache
