+++ modules/openapi-generator/src/main/resources/cpp-boost-beast-common/anytype-header.mustache
