+++ modules/openapi-generator/src/main/resources/cpp-boost-beast-common/licenseInfo.mustache
@@ -1,0 +1,11 @@
+/**
+ * {{{appName}}}
+ * {{{appDescription}}}
+ *
+ * {{#version}}The version of the OpenAPI document: {{{.}}}{{/version}}
+{{#infoEmail}} * Contact: {{{.}}}
+{{/infoEmail}} *
+ * NOTE: This class is auto generated by OpenAPI-Generator {{{generatorVersion}}}.
+ * https://openapi-generator.tech
+ * Do not edit the class manually.
+ */
