+++ modules/openapi-generator/src/main/resources/cpp-boost-beast-common/model-source.mustache
