+++ modules/openapi-generator/src/main/resources/cpp-boost-beast-common/oas31_deep_equal.mustache
