+++ modules/openapi-generator/src/main/resources/cpp-boost-beast-common/oas31_exact_number.mustache
