+++ modules/openapi-generator/src/main/resources/cpp-boost-beast-common/oas31_schema_ir.mustache
