+++ modules/openapi-generator/src/main/resources/cpp-boost-beast-common/validation-types.mustache
