+++ modules/openapi-generator/src/test/java/org/openapitools/codegen/cppboostbeast/Oas31ExactRuntimeTest.java
@@ -44,7 +44,7 @@ public class Oas31ExactRuntimeTest {
         Path includeDirectory = output;
         Path executable = output.resolve("oas31-exact-runtime-test");
         Path validationTemplate = Path.of(
-                "src/main/resources/cpp-boost-beast-client/validation-types.mustache");
+                "src/main/resources/cpp-boost-beast-common/validation-types.mustache");
         String modelNamespace = "org::openapitools::client::model";
         String validationNamespace = modelNamespace + "::detail::schema_validation";
         String validationGuard =
@@ -807,7 +807,7 @@ public class Oas31ExactRuntimeTest {
             Path output,
             String namespaceName,
             String guardPrefix) throws IOException {
-        Path templateDirectory = Path.of("src/main/resources/cpp-boost-beast-client");
+        Path templateDirectory = Path.of("src/main/resources/cpp-boost-beast-common");
         String[][] headers = {
             {"oas31_exact_number.mustache", "Oas31ExactNumber.h"},
             {"oas31_exact_json.mustache", "Oas31ExactJson.h"},
